Working on more advanced operator code.
Work required to support the member access operator ("." or "->").

There are some problems when it comes to Identifier evaluation. The
expression engine expects all of them to be variable references, which
is obviously not correct. Need a way to allow operators to to take the
decision on how to treat an identifier (as reference or actual
"keyword").