Moving forward with the evaluation
Directives can now be written and connected as expected to the
interpreter.

Working on simplifying the creation of directives by adding base classes
that can be easily extended by anyone.

Hiding a lot of internals as those are not really useful as they cannot
be extended in any practical matter. No point in exposing that much
detail to the consumer. Plus, removes the need to me to directly test
public APIs.