Going the "primitive" route.
Will try to catch as much "JS-like" behaviour as possible, as that makes
it at least a bit consistent. My code was too rigid in most cases, but I
was afraid to make it flexible fro the fear of unintended behaviour. Now
I think I found a good overall balance.